backend/requirements.txt
fastapi
uvicorn
joblib
pandas
python-dotenv
psycopg2-binary
